+++ src/utils/index.ts
@@ -5,6 +5,10 @@ export {default as getLength} from './getLength';
 export {default as getScrollParent} from './getScrollParent';
 export {default as getSpacing} from './getSpacing';
 export {default as scrollIntoViewIfNeeded} from './scrollIntoViewIfNeeded';
+export {
+	default as slicePathSegments,
+	stripSlashFromEnd,
+} from './slicePathSegments';
 export * from './spacing';
 export {default as truncateText} from './truncateText';
 export * from './units';